+++ dev-test/star-wars/__generated__/HeroTypeDependentAliasedField.tsx
@@ -9,11 +9,7 @@ export type HeroTypeDependentAliasedFieldQueryVariables = Types.Exact<{
 
 export type HeroTypeDependentAliasedFieldQuery = {
   __typename?: 'Query';
-  hero?:
-    | { __typename?: 'Droid'; property?: string | null | undefined }
-    | { __typename?: 'Human'; property?: string | null | undefined }
-    | null
-    | undefined;
+  hero?: { __typename?: 'Droid'; property?: string | null } | { __typename?: 'Human'; property?: string | null } | null;
 };
 
 export const HeroTypeDependentAliasedFieldDocument = gql`